move: 패키지를 AI 도메인으로 이동
