+++ src/main/java/com/nameplz/baedal/domain/AiRequest/service/AiRequestService.java
@@ -6,12 +6,12 @@ import com.nameplz.baedal.domain.AiRequest.domain.AiRequest;
 import com.nameplz.baedal.domain.AiRequest.dto.request.AiRequestDto;
 import com.nameplz.baedal.domain.AiRequest.dto.response.AiResponseDto;
 import com.nameplz.baedal.domain.AiRequest.mapper.AiRequestMapper;
+import com.nameplz.baedal.domain.AiRequest.property.AiProperty;
 import com.nameplz.baedal.domain.AiRequest.repository.AiRequestRepository;
 import com.nameplz.baedal.domain.user.domain.User;
 import com.nameplz.baedal.domain.user.repository.UserRepository;
 import com.nameplz.baedal.global.common.exception.GlobalException;
 import com.nameplz.baedal.global.common.response.ResultCase;
-import com.nameplz.baedal.global.config.AiProperty;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
